requirements.txt
fastapi
gunicorn
uvicorn
python-decouple
pytest
httpx
rav